core/avm1: Move AVM1-specific XML code into AVM1
