+++ core/src/avm1/xml.rs
@@ -1,0 +1,6 @@
+//! Garbage-collectable XML DOM impl
+
+mod iterators;
+mod tree;
+
+pub use tree::{XmlNode, ELEMENT_NODE, TEXT_NODE};
